Rename test for github mimetypes and move all test to it
